requirements.txt
PyQt5==5.15.10
google-generativeai==0.8.3
pydocstyle==6.3.0
matplotlib==3.7.1
